tests/wpt/tests/svg/animations/svgnumberoptionalnumber-animation-invalid-value-1.html

<!doctype html>
<title>Test SVGNumberOptionalNumber animation with invalid values: String values in from and to attributes.</title>
<link rel="help" href="https://www.w3.org/TR/2001/REC-smil-animation-20010904/#FromAttribute">
<link rel="help" href="https://www.w3.org/TR/2001/REC-smil-animation-20010904/#ToAttribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<svg id="svg">
    <defs>
        <filter>
            <feGaussianBlur id="blur" stdDeviation="5 5">
                <animate attributeName="stdDeviation" begin="0s" dur="4s" from="A Z" to="B X" />
            </feGaussianBlur>
        </filter>
    </defs>
    <rect x="50" y="50" width="100" height="100" fill="green" filter="url(#filter)"></rect>
</svg>

<script>

async_test(t => {
    const svg = document.getElementById("svg");
    const feGaussianBlur = document.getElementById("blur");

    window.addEventListener('load', t.step_func(() => {
        svg.setCurrentTime(2);

        requestAnimationFrame(t.step_func_done(() => {
          assert_equals(feGaussianBlur.stdDeviationX.baseVal, 5);
          assert_equals(feGaussianBlur.stdDeviationY.baseVal, 5);
          assert_equals(feGaussianBlur.stdDeviationX.animVal, feGaussianBlur.stdDeviationX.baseVal);
          assert_equals(feGaussianBlur.stdDeviationY.animVal, feGaussianBlur.stdDeviationY.baseVal);
        }));
    }));
  });
</script>
