LayoutTests/imported/w3c/web-platform-tests/svg/animations/svgnumberoptionalnumber-animation-invalid-value-1-expected.txt


FAIL Test SVGNumberOptionalNumber animation with invalid values: String values in from and to attributes. assert_equals: expected 5 but got 0
